dedupe create large recordset code
