+++ chromadb/test/property/test_add_mcmr.py
@@ -1,6 +1,5 @@
 import uuid
-from random import randint
-from typing import cast, List, Any, Dict, Tuple
+from typing import cast, List, Tuple
 import hypothesis
 import pytest
 import hypothesis.strategies as st
@@ -14,6 +13,7 @@ from chromadb.test.conftest import (
 )
 import chromadb.test.property.strategies as strategies
 import chromadb.test.property.invariants as invariants
+from chromadb.test.property.recordset_utils import create_large_recordset
 from chromadb.test.utils.wait_for_version_increase import wait_for_version_increase
 from chromadb.utils.batch_utils import create_batches
 from chromadb.api.client import AdminClient
@@ -229,29 +229,6 @@ def _test_add(
                 embedding_function=collection.embedding_function,
             )
 
-
-# Hypothesis struggles to generate large record sets so we explicitly create
-# a large record set
-def create_large_recordset(
-    dimension: int,
-    min_size: int = 45000,
-    max_size: int = 50000,
-) -> strategies.RecordSet:
-    size = randint(min_size, max_size)
-
-    ids = [str(uuid.uuid4()) for _ in range(size)]
-    metadatas = [{"some_key": f"{i}"} for i in range(size)]
-    documents = [f"Document {i}" for i in range(size)]
-    embeddings = [[1.0] * dimension for _ in range(size)]
-    record_set: Dict[str, List[Any]] = {
-        "ids": ids,
-        "embeddings": cast(Embeddings, embeddings),
-        "metadatas": metadatas,
-        "documents": documents,
-    }
-    return cast(strategies.RecordSet, record_set)
-
-
 @given(collection=collection_st, should_compact=st.booleans())
 @settings(deadline=None, max_examples=2)
 def test_add_large(
@@ -277,9 +254,9 @@ def test_add_large(
         )
 
     record_set = create_large_recordset(
-        dimension=collection.dimension,
         min_size=10000,
         max_size=20000,
+        embedding=[1.0] * collection.dimension,
     )
     coll1 = client1.create_collection(
         name=collection.name,
